backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
duckdb>=0.9.0
psutil>=5.9.0